requirements.txt
Flask
requests
Pillow
openpyxl
python-dotenv
transformers
torch
ollama
